requirements.txt
currency-converter-free
PyYAML
tabulate
pytest
